refactor(configure): extract capabilities subpackage, 15 -> 5 source files
The configure package held two components that shared nothing. Grepping every
symbol in the capabilities cluster (toolModeSpecs, modeParamSpec, toStringSlice,
inferDispatchParam, extractModes, buildParamDetails, buildModeParams) against every
symbol in the configure-tool cluster (SummarizeAuditEntries, ParseTestBoundary*,
Rewrite*Args) returns two file sets with zero overlap. describe_capabilities is not
even about configuring anything -- it introspects all five tools' schemas and happens
to be reachable as a configure mode. It now lives in
internal/tools/configure/capabilities (9 files), leaving configure at 5.

Eight of the fifteen files were minted by one splitting commit (58264122, "refactor:
split server internals into focused modules"), so the capabilities_*/mode_specs_*
prefixes are not independent evidence. Here the reference graph contradicts the
filenames in the opposite direction from the usual: the two prefixes look like two
groups but are one component, and they cannot be separated into two packages at all.
modeParamSpec flows capabilities -> mode_specs while toolModeSpecs flows
mode_specs -> capabilities, so any package boundary between them is an import cycle.
They ship as one package, and modeParamSpec moves into modespecs.go to sit with the
data it types.

Two merges inside the new package, both from the call graph:

- capabilities_schema.go + capabilities_param_details.go -> schema.go. Both do
  nothing but read raw JSON schema maps, and buildParamDetails decodes through
  toStringSlice from the other file. The package now has exactly one file that
  touches raw schemas.
- capabilities_mode_params.go -> capabilities.go. buildModeParams has a single
  caller, buildToolCapEntry, in the file it merged into.

The five per-tool spec files stay separate. That grouping is not an artifact: the
repo has exactly five MCP tools and the set is fixed, so one file per tool is a
stable one-to-one mapping, not an arbitrary slice of a 448-line table.

Response shapes are untouched -- BuildCapabilitiesMap/Summary/ForTool and
FilterToolByMode keep their names, signatures and every map key, since these back a
public MCP tool response. Exported symbol set is identical across the two packages
(15 symbols) and the multiset of non-comment code lines is identical at 990.
Combined coverage is 84.9%, exactly the pre-split configure figure, with no function
in the new package at 0.0%. One consumer changed:
cmd/browser-agent/internal/toolconfigure/capabilities.go, import plus four
qualifiers.
